requirements.txt
requests
azure-ai-textanalytics==5.3.0
python-dotenv